requirements.txt
pdfplumber>=0.10.0
pandas>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
